#659 fix tooltip z-index bug, block geo data
