+++ discovery-frontend/src/assets/css/metatron/popup/metatron.popup.02Workbook_ChartMap.css
@@ -81,6 +81,7 @@
 
 /* 라이브러리 tooltip */
 .ol-viewport {overflow:inherit !important;}
+.ol-viewport .ol-overlaycontainer .ol-overlay-container.ol-selectable {z-index:1;}
