+++ discovery-frontend/src/assets/i18n/ko.json
@@ -1806,6 +1806,7 @@
   "msg.storage.ui.list.geo.polygon" : "폴리곤",
   "msg.storage.ui.list.geo.block.custom.field.geo" : "GEO 데이터가 포함된 사용자필드는 사용할 수 없습니다",
   "msg.storage.ui.list.geo.block.custom.field.agg.function" : "Aggregate Function이 포함된 사용자 필드는 사용할 수 없습니다",
+  "msg.storage.ui.list.geo.block.other.charts" : "GEO 데이터는 맵뷰에서만 사용할 수 있습니다",
   "msg.storage.ui.list.expression" : "표현식",
   "msg.storage.ui.list.type" : "타입 변경",
   "msg.storage.ui.list.del" : "삭제",
